chore: remove meteor.startup from pin-message
