+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -4,7 +4,6 @@ import './jumpToSearchMessage';
 import './jumpToStarMessage';
 import './permalinkPinned';
 import './permalinkStar';
-import './pinMessage';
 import './starMessage';
 import './unpinMessage';
 import './unstarMessage';
